+++ lab6/CMakeLists.txt
@@ -3,4 +3,4 @@ project(lab6)
 add_subdirectory_if_exists(counter)
 add_subdirectory_if_exists(academia)
 add_subdirectory_if_exists(zipper)
-add_subdirectory_if_exists(pointt)
+add_subdirectory_if_exists(point)
